fix: split mux and executor
